Correct license headers introduced by UPB
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
from BSD license headers that got introduced by mistake
by Vlad-Andrei BĂDOIU with the following commits:
 d2967b3b1 - lib/ukswrand: Adapt the library to work with multiple algorithms
 9bb9a617d - lib/ukswrand: Add ChaCha algorithm
 8c5e2b4bb - lib/ukswrand: Add /dev/random and /dev/urandom
 2784ae201 - lib/nolibc: Add pwd.h
 19ebca61d - include/uk: Add page.h
 8c9cd0ec9 - lib/uksysinfo: Introduce uksysinfo library
 3a39d50dd - plat/*: Introduce unikraft internal constructors
 200037918 - lib/nolibc: Add utsname.h

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

The original author consents to the change with a `Signed-off-by`
or `Reviewed-by` tag to this commit.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro> on behalf of Vlad-Andrei BĂDOIU