Correct license headers introduced by UPB
This commit removes the additional clause
"THIS HEADER MAY NOT BE EXTRACTED OR MODIFIED IN ANY WAY"
from BSD license headers that got introduced by mistake
Roxana Nicolescu with the following commit:
 cd1873671 - lib/ukblkdev: Request interface
 9065c15ef - plat/xen/drivers/blk: Register blkfront with libukblkdev
 30f42955f - plat/xen/drivers: Introduce blkfront driver skeleton
 620425c02 - lib/ukblkdev: Blkdev registration
 bff097cd2 - plat/xen/drivers/blk: Init Xenbus Device

On the one hand, this additional clause is redundant because
the BSD license already states that it must remain and the
copyright notice must be kept.
On the other hand, the clause freezes the file header
and prohibits future contributors from extending the
copyright notice for their contributions. This additional
clause is not part of the official BSD 3-clause.

The original author consents to the change with a `Signed-off-by`
or `Reviewed-by` tag to this commit.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Signed-off-by: Roxana Nicolescu <nicolescu.roxana1996@gmail.com>