build: fetch: Support checksum checks
This commit introduces the support for comparing the checksum of a
downloaded archive with `fetch`, `fetch2`, `fetchas`, and `fetchas2`.
It computes and compares the checksum with the library-local variables
within a `Makefile.uk` when given:
 LIBNAME_ORIGIN_[MD5|SHA1|...].
Please note that this variable needs to be defined
before calling `fetch`, `fetch2`, `fetchas`, or `fetchas2`.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>