support/build: Extend preprules for external language support
We extend the build system to load preparation rules from external
libraries. This enables Unikraft to support preparation rules for
different formats without modifying the core Makefile.rules.

Usage: If we want to add a new external preparation rule, we
create in the root of the library a  Makefile.rules. Inside
Makefile.rules we define a preprule_{format}.

The preprule should be similar to the preprule_m4 defined in
support/build/Makefile.rules.

Signed-off-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@upb.ro>