arm/entry: set x0 before cache maintenance.
There is a long standing issue in entry code of arm that x0 is not
set before using it in clean_and_invalidate_dcache_range. This error
can be caught by kvm and result in core dump.
Here, x0 is set to the location where stores dtb before cache maintain.

Signed-off-by: Jianyong Wu <jianyong.wu@arm.com>
Reviewed-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Approved-by: Michalis Pappas <mpappas@fastmail.fm>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #317