lib/uktest: Provide multi boot-level registration of test suites
This commit introduces a minimal implementation testing
infastructure into Unikraft.  The primary contents includes the
definition of a test (or suite of tests) and the mechanism by
which one can use to register the suite to be run.

In `uktest`, tests are organised hierarchically powered by the
the lowest common denominator: the assertion.  This organisation
is inspired by KUnit[0], the Linux Kernel's in-house testing
system.  For licensing reasons the Unikraft project cannot use
this source code.  However, by inspiration, we can organise the
`uktest` library following a similar pattern:

 1. The assertion: repersents the lowest common denominator of a
    test: some boolean operation which, when true, a single test
    passes.  Assertions are often used in-line and their usage
    should be no different to the traditional use of the `ASSERT`
    macro.  In this commit, we introduce a new definition:
    `UK_TEST_ASSERT` which has two parameters: a reference to the
    test case (see 2.) and the boolean opeation which is
    true-to-form of the traditional `ASSERT` macro: some boolean
    expresson.  This commit extends this functionality to simply
    note when this assertion has failed and to print
    contextualised information about it.

 2. The test case: often, assertions are not alone in their means
    to check the legitimacy of operation in some function.  We
    find that a "case" is best way to organise a group of
    assertions in which one particular function of some system is
    under-going testing.  A case is independent of other cases,
    but related in the same sub-system.  For this reason we
    register them together.

 3. The test suite: represents a group of test cases.  This is
    the final and upper-most heirarchical repesentation of tests
    and their groupings.  With assertions grouped into test cases
    and test cases grouped into a test suite, we end this
    organisation in a fashion which allows us to follow a common
    design pattern within Unikraft: the registration model.  The
    syntax follows similar to other registation models within
    Unikraft, e.g. `ukbus`.  However, `uktest`'s registation
    model is more powerful (see Test Execution).

To register a test suite with `uktest`, we simply invoke
`uk_testsuite_register` with an initialized `struct uk_testsuite`
variable.  This structure contains an initialized list of `struct
uk_testcase`s.  Each test case has only one input parameter: a
reference to the current case which each assertion handles during
its conditional check:

```c
UK_TESTCASE(testsuite1, some_test_case)
{
	int x = 1;
	UK_TEST_ASSERT(x > 0);
}

uk_testsuite_register(testsuite1, NULL);
```

The above snippet can be organised within a library in a number
of ways such as in-line or as an individual file representing the
suite.  There are a number of test suite registration handles
which are elaborated on in next section.  It should be noted that
multiple test suites can exist within a library in order to test
multiple features or components of said library.

In order to achieve consistency in the use of `uktest` across the
Unikraft code base, the following recommendation is made
regarding the registration of test suites:

 1. A single test suite should be organised into its own file,
    prefixed with `test_`, e.g. `test_feature.c`. All tests
    suites of some library within Unikraft should be stored
    within a new folder located at the root of the library named
    `tests/`.
 2. All tests suites should have a corresponding KConfig option,
    prefixed with the library name and then the word "TEST", e.g.
    `LIBNAME_TEST_`.
 3. Every library implementing one or more suite of tests must
    have a new menuconfig housing all test suite options under
    the name `LIBNAME_TEST`.  This menuconfig option must invoke
    all the suites if `LIBUKTEST_ALL` is set to true.

`uktest`'s registation model allows for the execution of tests
at different levels of the boot process.  All tests occur before
the invocation of the application's `main` method.  This is done
such that the validity of the kernel-space functions can be
legitimised before actual application code is invoked.  A fail-
fast option is provided in order to crash the kernel in case of
failures for earlier error diagnosis.

When registering a test suite, one can hook into either the
constructor "`ctor`" table or initialisation table "`inittab`".
This allows for running tests before or after certain libraries
or sub-systems are invoked during the boot process.

The following registation methods are available from this commit:

 * `uk_testsuite_early_prio`,
 * `uk_testsuite_plat_prio`,
 * `uk_testsuite_lib_prio`,
 * `uk_testsuite_rootfs_prio`,
 * `uk_testsuite_sys_prio`,
 * `uk_testsuite_late_prio` (default),
 * `uk_testsuite_prio`, and
 * `uk_testsuite_register`.

[0]: https://kunit.dev

Checkpatch-Ignore: MACRO_WITH_FLOW_CONTROL
Checkpatch-Ignore: COMPLEX_MACRO
Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>