plat/drivers: Add KVM balloon implementation
The implementation of the KVM balloon driver is based on other already
added Virtio drivers from the project. It follows a similar structure
to them. After adding the device, the driver uses virtqueues to
send and take memory through the hypervisor.

The system will interact with the driver through the calls exposed
in the header.

Signed-off-by: Jack Raney <raney.jack99@gmail.com>
Signed-off-by: Cason Schindler <cason.j.schindler@gmail.com>
Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>