plat/xen: Add Xen balloon hypervisor calls
The Xen implementation just prepares the arguments to call the Xen
hypervisor.

As Xen reserves and keeps separate memory zones for it's VM's, it is
harder to observe the behaviour of the balloon. To add to this, Xen
already has a working balloon that takes memory from the host, which
makes things a bit harder to test.

Signed-off-by: Jack Raney <raney.jack99@gmail.com>
Signed-off-by: Cason Schindler <cason.j.schindler@gmail.com>
Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>