lib/ukallocbbuddy: Insert balloon API calls
In order to work with the balloon, the API calls were added to bbuddy.
The `palloc` function calls deflate, `pfree` and `addmem` call inflate.

The balloon is first started using the `balloon_init` function that is
called just once. In the case of KVM, because it does not currently
need to initialize itself by inflating the balloon, the calls to
`balloon_init` and `ukplat_inflate` are skipped.

Signed-off-by: Jack Raney <raney.jack99@gmail.com>
Signed-off-by: Cason Schindler <cason.j.schindler@gmail.com>
Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>