arch/x86: Add trap interface headers
The event-based trap interface makes architectural traps accessible as
events (uk_event). This commit adds the header files for x86, which
contain the available traps and the context data supplied to trap
handlers. A trap handler for one of the TRAP_* events may be
registered using the uk_event API.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>