arch/arm64: Add trap interface headers
The event-based trap interface makes architectural traps accessible as
events (uk_event). This commit adds the header files for arm64, which
contain the available traps and the context data supplied to trap
handlers. A trap handler for one of the TRAP_* events may be
registered using the uk_event API.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Michalis Pappas <mpappas@fastmail.fm>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #231