arch/arm64, plat/*: Change __regs and fix trap
This commit does multiple things:
1) Removes the EL0 stack pointer from the arm64 __regs structure.
There is no meaningful use for the EL0 stack pointer as all code
runs in EL1. Even if we were entering from EL0, the new trap code
saves EL0_SP in the shared sp field. As there is no alignment of EL1_SP
when coming from EL0, EL1_SP can be restored simply by subtracting
__REGS_SIZEOF (i.e., there is no need to explicitly store EL_SP1 in
__regs). That also makes it more intuitive to use the sp field in __regs
(e.g., in crash code, debuggers, etc.), as it always represents the
stack pointer of the trapped code, no matter if this was running in
EL0 or EL1. In addition, in the old code SP1 would erroneously
point below the trap frame (e.g., this breaks stack traces).
2) Removes the ESR_EL1 field from the arm64 __regs structure and
supplies it via an extra parameter to the trap functions.
There is no use case for restoring ESR_EL1 when leaving the trap, it
can thus be removed, which also alignes the __regs structure to
16-bytes again, after removing SP0.
3) Fixes a bug, where x19 is not being restored after a trap.
4) Fixes a bug, where EL1_SP is incorrectly restored, leaking
memory, if it was unaligned at trap entrance.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>