lib/ukstore: Introduce library skeleton
The `ukstore` library has a dependency to `nolib`/`newlib` and to
`ukdebug`. As such they need to be selected to work.

The symbols exported will be provided by the `store.c` file.

The Makefile contains rules to build all files and generate files.
The Makefile first generates a list of all available libraries and
then generates two headers and a linker script to create and access
sections for each available library.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>