lib/ukstore: Add static entry registration
This commit introduces a method to allocate static elements in library
sections. Each library has its own section generated at build.
This type of entries do not take up time and are not modified.
Also, these entries do not use folders (the folder is the section).

These will be used to pass system information getters/setters
throughout the system in an easy-enough manner. Interaction with this
type of entries is done through the same API as the dynamic ones.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>