plat/common: Introduce page tables API
Here are the functions used to manipulate memory mappings, agnostic of
architecture and platform. Since the strucutre of the page tables is
hierarchical with 4 levels on both x86_64 and ARM64, the logic for page
table walking is the same.

Differences are in the abstracted ukarch_* functions, which were
introduced in the previous patches for the x86_64 architecture.

uk_pt_init() is the function that should first be called to initialize
the API. It keeps the existing 1:1 mappings until a physical address,
given as a parameter. Then, it initializes the internal bookkeeping of
the API (the frame allocator) and returns a uk_pagetable instance.

uk_pt_build() is the function used for building new page tables (that
are used in parallel with the first one returned by uk_pt_init). This
allows Unikraft to have multiple address spaces at the same time.

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>