plat/common: Add map/unmap functions to page table API
Introduce functions for creating new mappings (for one page at a time or
for multiple), for unmapping existing pages. These functions also have
some internal versions (named with _initmem) that are used during the
initialization of a new page table, when we don't write to the active
page table (so we don't use the platform provided mechanisms of
modifying the page table - like hypercalls, on Xen - but we just write
to a normal memory location).

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>