plat/kvm: Enable page table API and dynamic initialization
If the page tables API configuration is selected, the memory layout defined
in mem_layout.h is used. If POSIX mmap functions are enabled, the heap memory
is divided in half between the heap and the memory available for mmap.
This is a temporary solution, and should be replaced by implementing a
mechanism for dynamically changing the heap size (e.g. brk()).

Signed-off-by: Stefan Teodorescu <stefanl.teodorescu@gmail.com>