lib/uktest: Add methods for retrieving stats about tests
This commit introduces minimal statistic retrieval within
`uktest` such that we can count the number of failed assertions,
cases and suites.  This commit also introduces a mechanism to
print these statistics at the last possible moment from the
`inittab` and thus the last possible moment a suite can be
registered.

Signed-off-by: Alexander Jung <a.jung@lancs.ac.uk>