lib/uksched: Add locking to waitq wait event
This commit does two things:
First, it adds the ability to supply an optional lock to the wait event
macro. The lock is released right before yielding execution to another
thread and acquired again as the first operation when woken up.
Thus unlocking/locking the mutex and blocking/unblocking the thread
become atomic operations from the perspective of the caller. As the
condition is tested under the lock, we essentially have a condition
variable, thus preventing signal loss with SMP.

If all callers of wait event would provide a lock and the thread
manipulations would be thread-safe the IRQ-protection could be
removed.

Second, the commit makes the wait event macro return if the deadline
has been hit so that the caller does not need to recheck the deadline
condition.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: <cristianvijelie@gmail.com>
Approved-by: Hugo Lefeuvre <hugo.lefeuvre@manchester.ac.uk>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #288