include/uk/arch/spinlock.h: Rewrite spinlock header
Allow custom architecture dependent spinlock implementation on multicore
systems through uk/asm/spinlock.h header file.

On single core systems spinlocks functions are void operations. No state is
saved for the spinlock but we assume that the CPU is always able to get the
lock. Trying to acquire a spinlock thus always succeeds and asserting that
a lock is held returns false.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Signed-off-by: Renê de Souza Pinto <Rene.deSouzaPinto@opensynergy.com>
Reviewed-by: Michalis Pappas <michalis.pappas@opensynergy.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #289