arch/x86_64: increase TLS alignment
Some CPU instructions require the operand to be aligned on a 16-byte
or 32-byte boundary. Such operations may involved thread-local
variables. When it is the case, the compiler (and linker) will increase
the alignment requirements for the `.tdata` and `.tbss` sections in the
ELF objects, and in TLS program header.

As Unikraft `uksched` TLS handling currently ignores alignment
requirements set in the ELF image, this commit increases the constant
alignment from 8 to 32, to allow for all use cases.

Signed-off-by: Cyril Soldani <cyril.soldani@uliege.be>
Reviewed-by: Sergiu Moga <sergiu.moga@protonmail.com>
Approved-by: Vlad-Andrei Badoiu <vlad_andrei.badoiu@upb.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #287