drivers/virtio-net: Do not print hardware address
Do not print the hardware address during feature negotiation.
For the case that the hardware address is overwritten or is
not used by an actual Unikraft application or network stack,
this information might be misleading for users.
Suggestion is to print the address by the application or
network stack.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>
Approved-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #308