arch/x86_64: add dummy thread-local variable
This ensures there will be at least one variable in the `.tbss` section.
If that section is empty and only contains padding for alignment
purpose, then TLS offsets may be wrong.

Closes #290.

Note that is would not be necessary if there was no padding added to the
`.tbss` section by the linker script, but x86_64 TLS code would then
have to be updated to read the actual required alignment from the ELF
file. Using a fixed alignment is much simpler to implement, and the
introduction of this thread-local variable should only cost 4 additional
bytes per thread.

Signed-off-by: Cyril Soldani <cyril.soldani@uliege.be>