lib/*: Fix gcc-11 warnings/errors regarding time
In gcc-11 mismatches between pointer formats are treated as warnings.
This is generated by the `timeval` structure:
`const struct timeval times[2]` vs. `const struct timeval *times`

In the case of Unikraft these warnings are treated as errors and the
build fails.

This commit fixes all the new mismatches by switching to the second
version. This should not change behaviour.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>