build: Restore silent build
With the introduction of verbose level 2 (`make V=2`), the
build variables `$(KBUILD_VERBOSE)` and `$(Q)` were not in
sync anymore. This caused a partial verbose build even if
`make V=0` was specified.
Additionally, this commit uses `verbose_cmd` for the kconfig
tools rule for a better seamless integration regarding
verbosity.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request: #323