lib/uksignal: Update struct sigset_t and macros to 1024bits
Musl uses 1024 bits for the signal mask, Newlib uses only 64
bits. Although there is no need for 1024 signals we still
need to adapt the `sigset_t` type to 1024 bits for musl
because otherwise the build will fail. This commit
redefines macros such as `sigaddset` to work with
an array rather than an unsigned long.

Signed-off-by: Dragos Iulian Argint <dragosargint21@gmail.com>