Order dependencies for prepare and preprocess targets
The `prepare` target in the Makefile depends on the `fetch` target.
The dependency is marked as an order-only dependency (`| fetch`), that
is run after all other dependencies.

This causes an issue when the `$(UK_PREPARE-y)` prerequisite depends on
the `fetch` target. This is the case with the `app-micropython`
application, where the `prepare` target for `lib-micropython` relies on
the `lib-newlib` library being fetched beforehand.

This commit solves this by adding the `fetch` target as a prerequisite
for `prepare` **before** the `$(UK_PREPARE-y)` prerequisite.

For consistency, the `prepare` target is added as a prerequisite for
`preprocess` **before** the `$(UK_PREPROCESS-y)` prerequisite.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>