plat/kvm/x86: Add paging to platform setup
Extends the platform setup code to enable paging. The code transforms
the existing static boot page table into a dynamic one that can be
changed by the paging API. The setup takes away the RAM assigned to
the original heaps and gives it to the frame allocator. In addition, all
further RAM regions supplied in the multiboot information are added
to the frame allocator, allowing to map/access all physical memory in
the machine.

The setup code is limited in that there is no dynamic virtual address
space management or demand paging. Therefore, all available RAM is
mapped as heap. Furthermore, the heap base address is fixed at
0x400000000.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Sergiu Moga <sergiu.moga@protonmail.com>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Approved-by: Michalis Pappas <michalis.pappas@opensynergy.com>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #338