lib/ukstore: Introduce library skeleton
The `ukstore` library has a dependency to `nolib`/`newlib` and to
`ukdebug`. As such they need to be selected to work.

The symbols exported will be provided by the `store.c` file.

The Makefile contains rules to build all files and generate files.
The Makefile first generates a list of all available libraries and
then generates two headers and a linker script to create and access
sections for each available library.

Signed-off-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>
Reviewed-by: Andrei Mutu <andrei.mutu@stud.acs.upb.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Approved-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #459