lib/ukmmap: Only unmap whole regions
The current implementation of munmap tries to unmap parts of
a memory region by allocating a new smaller region and copying the
remaining memory contents to the new area. However, this changes
the base address of the memory area, which cannot be communicated
to the caller. The behavior is thus broken. This commit only deallocates
memory areas if a whole VMA is to be unmap. Otherwise, we cannot
do anything about it and ignore the call.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Hugo Lefeuvre <hugo.lefeuvre@manchester.ac.uk>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #394