lib/uktest: Fix write protection of testtab on ARM
Due to is position behind the `.text` section, the current platform
initialization code that sets up the page tables marks the testtab
as read-only. However, we update the assertion status whenever we
performed a test. This leads to a fatal crash. This commit moves the
table so that it is no longer affected by the write-protection.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>