docs: Migrate docs and update references to contribution guide
This commit removes the (now very outdated) `docs/` folder from the
repository and `CODING_STYLE.md` file.  Towards a greater effort of keeping
the docs up-to-date and in-line with the growing Unikraft ecosystem, the
location of these documents have migrated to:

  https://github.com/unikraft/docs.git

This standalone repository consolidates documentation across the project,
including the core of which docs used to be located.

References to the documentation and how to contribute are updated in this
commit to reflect this change.

Signed-off-by: Alexander Jung <alex@unikraft.org>
Reviewed-by: Gabi Mocanu <gabi.mocanu98@gmail.com>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #252