lib/posix-socket: Add POSIX socket driver library
This library allows for the registration of socket drivers that
implement the posix_socket_ops interface and thereby expose a
POSIX-compliant socket API for a certain AF address family. Drivers
may register as many AF numbers as desired or specify new families
which are not listed in the <sys/socket.h> header. To access the
relevant implementation, the application simply creates a socket
with the required AF family number.

Signed-off-by: Alexander Jung <alex@unikraft.io>
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #65