lib/vfscore, lib/*fs: Use non-prototyped no-ops
VFS operations are implemented with function calls. In order to stub
some of the operations a no-op function is assigned to those function
pointers. As the function signature between the generic no-op function
and the respective VOP does not match, the compiler displays a
warning. We currently suppress this warning with a compile flag.
However, this also suppresses legitimate warnings and may hide bugs.
This commit changes the no-op functions to be non-prototypical so
that any function signature is treated as valid. The compile flags can
thus be removed.

Checkpatch-Ignore: FUNCTION_WITHOUT_ARGS
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>