lib/vfscore: Fix return value of __xstat
The stat syscall is defined as raw syscall and directly returns
the return value of __xstat. However, __xstat returns -1 and
sets errno which is supposed to happen in the syscall wrapper.
This commit changes the return value to the negative errno
value. Although, this breaks direct calls to __xstat this is inline
with our current implementation of __lxstat, __statfs, __fstatfs,
and others.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>
Approved-by: Simon Kuenzer <simon@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #411