lib/vfscore: Fix return values of read/write
In case of a read/write error, the read/write syscall implementations
should return the negative errno. However, the current implementations
call the C-library exports which set the errno and return -1. This
commit fixes the functions to call the raw syscall implementations
directly just like already done for pread, for example.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>