lib/ukdebug: Option to print stack frame and caller address
Replaces the stack print option with printing the current stack frame
address and the address that called a kernel or debug print. This is a
preparation for supporting custom sized stacks where the base address of
a stack cannot be computed anymore from the current stack pointer
address.

Checkpatch-Ignore: COMPLEX_MACRO
Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Gabi Mocanu <gabi.mocanu98@gmail.com>
Reviewed-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #524