plat/linuxu: Add support for arm64 on linuxu
This commit adds support for arm64 architecture on linuxu platform.
It adds all the necessary syscalls for arm64 in a new header file.
Also, I added the necessary changes in the syscall.h and Makefile.uk files
so that it will compile with arm64 syscalls when that option is selected.

Signed-off-by: maniatro111 <florin.postolache80@gmail.com>
Reviewed-by: Stefan Jumarea <stefanjumarea02@gmail.com>
Approved-by: Alexander Jung <alex@unikraft.org>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #418