arch/arm/arm64: plat/common/arm: Add Pointer Authentication support
Pointer Authentication (FEAT_PAuth) allows signing and authenticating
pointers to harden systems against classes of attacks that rely on
memory corruption, such as stack smashing and ROP.

Pointer Authentication Codes (PACs) are created by signing a pointer and
a 64-bit modifier with an 128-bit key. The modifier is a value that is
used to restrict the PAC to a specific context (eg the SP). Keys are
stored in registers. The PAC is stored in the unused upper bits of the
memory address, and can be later verified to ensure that the pointer has
not been tampered with.

This commit adds basic Pointer Authentication support for platforms that
support arm64. Specifically, it introduces:

- The CONFIG_ARM64_FEAT_PAUTH option. When enabled, GCC is instructed
  to generate PACIASP/RETAA sequences to protect the return address of
  functions using the PACIA_Key.

- An API to set up and enable Pointer Authentication. Platforms are
  required to implement a key generation function that uses a secure
  source of randomness.

- A macro to disable Pointer Authentication in certain functions. This
  is required on any returning function that is called prior to the
  initialization of Pointer Authentication.

Unikraft images generated with FEAT_PAuth are compatible with
architectures that implement Armv8.3 and above. That is due to the fact
that, although the basic set of Pointer Authentication instructions is
implemented in the NOP space, combined instructions and registers
holding the keys are not.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@gmail.com>
Approved-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #369