lib/syscall_shim: Filter provided syscalls based on architecture
Currently, when we register a syscall to the syscall shim layer,
we don't check if it should be available on the target architecture.
Therefore, several build errors appear when some libraries are selected.
This commit adds an additional intermediary file in the build process,
which helps registering a syscall only if it is present in the
corresponding `syscall.h.in` architecture file.

Signed-off-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Signed-off-by: Răzvan Deaconescu <razvan.deaconescu@cs.pub.ro>