support/build: Prefer local include dirs
For include directories the preprocessor searches the paths in
left-to-right order. In case of conflicting "system" headers, the
previous order caused a surprising behaviour: the system headers would
be preferred to the library local headers. Using ~#include "header.h"~
would be the proper solution, but this can be inconvenient when
packaging external libraries for unikraft.

Signed-off-by: Marco Schlumpp <marco.schlumpp@gmail.com>