lib/syscall-shim: Save & restore ECTX on binary system calls
Compiles the binary system call handler code as ISR-safe. Additionally,
the extended context is saved and restored of the system call caller on
each binary system call. The reason is that extended register sets maybe
used by the system call implementations and the Linux system call
calling convention assumes that caller programs do not have to save the
state themselves.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>