lib/uksched: Fix race condition in wait_event
__wq_wait_event_deadline() waits until a certain condition is met or
a deadline is reached. The current code does a preliminary condition
check to avoid disabling interrupts just to find that the condition
evaluates to true. However, if the condition is actually false, it is
not checked again after disabling interrupts. If the condition
changes in an interrupt handler this can lead to a signal loss. The
thread then goes to sleep although it should continue to run.

The commit adds the additional check and bails out if the
condition is met before blocking the thread.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Reviewed-by: Simon Kuenzer <simon@unikraft.io>
Approved-by: Simon Kuenzer <simon@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Pull-Request-Closes: #427