lib/ukargparse: Fix quoted whitespaces handling
Previously, the whitespaces were not handled when
placed inside a quoted text and at the beginning
of a new word (i.e when `prev_wspace` == 1 and
`in_quote` != 0).
This led to arguments like `"\targ"` be parsed
to `arg` instead of `\targ`.

This commit treats the whitespaces when inside a
quoted text the same as all the other characters
in the `default` branch of the switch statement,
so when `prev_wspace` is set to 1 a new entry is
created in `argv` and `prev_wspace` is set to 0.

Signed-off-by: Stefan Jumarea <stefanjumarea02@gmail.com>