support/scripts: Enable generation of GRUB ISO images
This support script allows easy generation of GRUB ISO images from
Unikraft KVM images. This script is intended and designed to be
expanded to allow more formats and architectures. This script is
designed to resolve #342 as it was determined this functionality would
be better placed outside of the Unikraft build pipeline.

Signed-off-by: Aadhithya Kannan <aadhi0319@gmail.com>
Reviewed-by: Andrei Mutu <andrei.mutu@stud.acs.upb.ro>
Reviewed-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Approved-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #342