lib/uksched: Provide thread container creation
There are use cases where a thread needs to be created before
it can be initialized with a context. For this purpose,
this commit introduces `uk_thread_create_container()` and
`uk_thread_create_container2()`. The first function takes care
of allocating a stack and TLS, while the second variant expects
a stack and TLS to be given by the caller.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #564