lib/uksched: Provide thread container creation
There are use cases where a thread needs to be created before
it can be initialized with a context. For this purpose,
this commit introduces `uk_thread_create_container()` and
`uk_thread_create_container2()`. The first function takes care
of allocating a stack and TLS, while the second variant expects
a stack and TLS to be given by the caller.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>