lib/uksched: Replace `uk_sched_get_default()` with `uk_sched_current()`
Instead of maintaining a default scheduler this commit provides a
function that returns the current scheduler of the current thread.
The original intention was to have a fall-back scheduler to which a
newly created thread should be assigned to. Using the current scheduler
as the fall-back policy is more appropriate when multiple schedulers are
used in the system.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #564