lib/uksched: Custom TCB initialization hooks
This commit introduces the possibility for a third-party threading
library to initialize a custom thread control block (TCB). The layout
of the TLS and TCB is defined by the compiler of the target architecture
and of which Unikraft only utilizes the TLS. As soon as a library
(e.g., pthread-embedded, musl) activates the invisible configuration
option `CONFIG_LIBUKSCHED_TCB_INIT`, the following two
hooks are called during thread creation and removal:
- `uk_thread_uktcb_init()`: On TLS allocation during thread creation
- `uk_thread_uktcb_fini()`: On TLS removal of a thread
These hooks are called by any `lib/uksched` API call that create or
terminate a thread so that the uksched API seamlessly integrates into
the third party threading library.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #564