lib/ukalloc: Add basic memory tagging support
This commit updates ukalloc with basic support for memory tagging.

Specifically it adds a new `size` field to ifpages metadata in order
to track the requested allocation size.

uk_malloc_ifpages(), and uk_realloc_freepages() are instrumented to tag
the requested region and return a tagged pointer. Tagging is applied to
the size requested by the caller (aligned to the tag granule) excluding
metadata and additionally allocated memory to page size.

uk_free_ifpages() is updated to tag the freed region with a new tag,
to prevent use-after-free.

The implementation is simple, as it does not implement optimizations
based on MTE features such as store allocation with zeroing, store
allocation tags for blocks, etc.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Approved-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #458