plat/common/arm: Add PAuth / BTI instructions to SMCCC conduits
Update assembly implementation of SMCCC conduits with the necessary
PAuth and BTI instructions. This fixes an issue where any code that
invokes an SMCCC conduit will cause a Data Abort, due to the lack
of branch protection instructions. That will bring the system into
an infinite loop, as `system_off()` invokes a PSCI call, yet again
via an SMCCC conduit.

Signed-off-by: Michalis Pappas <mpappas@fastmail.fm>