lib/uksched: Thread creation helper functions
This commit introduces thread creation helper functions that create a
thread that is directly assigned to the current scheduler. For the
creation default allocators are used that are set by the scheduler.
Variants for the most common cases are provided:
- `uk_sched_thread_create_fn0()`: thread function without any argument
- `uk_sched_thread_create_fn1()`: thread function with one argument
- `uk_sched_thread_create_fn2()`: thread function with two arguments
In addition, the shortcut macro `uk_sched_thread_create()` is provided
for a convenient thread creation with just four arguments: (1) the
scheduler instance, (2) the thread function to execute (one argument),
(3) an argument pointer, and (4) an optional thread name.
This macro basically replaces the previous helper `uk_thread_create`()`.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Reviewed-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Cezar Craciunoiu <cezar.craciunoiu@unikraft.io>
Reviewed-by: Adina Smeu <adina.smeu@gmail.com>
Reviewed-by: Razvan Deaconescu <razvand@unikraft.io>
Approved-by: Razvan Deaconescu <razvand@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #564