lib/uksignal: Solve redefining conflicts
There are many functions that are defined in musl as well as in uksignal.
When using musl we want to use the musl functions. When we use musl the
`LIBSYSCALL_SHIM_NOWRAPPER` macro is set, which leads to the setting of
`UK_LIBC_SYSCALLS` to 0. We can use this to exclude the redefined
functions from vfscore.

Signed-off-by: Dragos Iulian Argint <dragosargint21@gmail.com>
Reviewed-by: Sergiu Moga <sergiu.moga@protonmail.com>
Approved-by: Simon Kuenzer <simon.kuenzer@neclab.eu>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #453