lib/uksched: Call termination callbacks on thread exit
The thread termination callbacks were called only when a thread was
released. Especially for self-terminating threads this meant that the
termination callback was not called until the garbage collector
released the thread resources.
This commit implements the expected behavior, namely that the thread
termination callbacks are called when the thread transitions to the
exited state. Whenever `uk_thread_release()` is called on a non-exited
thread, the state transition to the exited state is enforced first. This
ensures that termination callbacks are called even in such a case.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>