lib/posix-info: Comment out unused sysinfo buffer
The sysinfo structure defines the `reserved` buffer. This is however not
present on Linux, meaning that binary-compatibile applications use a
shorter buffer. This results in a buffer overflow when transferring
information from Unikraft to the buffer in binary-compatible app.

This commit fixes this by commenting out the `reserved` buffer in the
`sysinfo` structure.

Signed-off-by: Razvan Deaconescu <razvan.deaconescu@cs.pub.ro>
Reviewed-by: Radu Nichita <radunichita99@gmail.com>
Approved-by: Marc Rittinghaus <marc.rittinghaus@unikraft.io>
Tested-by: Unikraft CI <monkey@unikraft.io>
GitHub-Closes: #673