lib/syscall_shim: Warn and ignore unnumbered system calls
Adds a warning issued by the pre-processor for system calls that
do not have a system call number defined. This happens when
a system call was registered to `UK_PROVIDED_SYSCALLS-y` but
does not have a number mapping in `syscall.h.in`.
The auto-generated header `provided_syscalls.h` does not
provide prototype definitions and the syscall_shim will
not link to ignored system calls.

Signed-off-by: Simon Kuenzer <simon.kuenzer@neclab.eu>