uk/plat: Change SMP API to use CPU index
The current SMP API assumes that logical CPUs are always sequentially
numbered starting from 0 up to the number of CPUs in the system.
However, this is not the case, for example, on native ARM64 and x86-64
machines. This commit thus introduces `__lcpuidx` which provides the
user with the sequential numbering scheme whereas the SMP code
internally references the actual CPUs with `__lcpuid`-typed IDs.
Since physical IDs often encode the CPU topology, we use a 64-bit type.

Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>