plat/common: Add generic implementation of SMP API
This commit adds an implementation for the SMP API defined by the
`ukplat_lcpu_(id|idx|count|start|run|wait|wakeup)()` functions and adds
signatures for the architectural-dependent parts that are called from
the generic functions.

To represent a logical CPU in the system we define the `lcpu`
structure which is stored in a fixed-sized array. To avoid cache
coherency overhead `lcpu` should be a multiple of the cache line size.

All `lcpu`s must be allocated from the array by the BSP before
starting secondary cores (BSP must be the first LCPU) as working
on the `lcpu` array is not synchronized. This should be done as
part of `lcpu_arch_mp_init()`. Every logical CPU in turn must jump
to `lcpu_entry_default` when leaving the boot trampoline. This
function is responsible for setting up the LCPU using lcpu_init()
and for jumping to the user-supplied entry function if existing.

In this commit, translating an LCPU ID to an LCPU index is rather
expensive as it involves a linear O(n) search over all allocated
`lcpu`s. A future commit should replace this with a faster algorithm.
If paging is enabled, `lcpu` structs could be allocated from physical
memory and mapped as indiviual pages with an additional CPU-local
mapping of the corresponding LCPU's struct mapped at a known
fixed address (i.e., every CPU has its own PT), replacing the
translation with a simple lookup of `lcpu->idx`.

Checkpatch-Ignore: WEAK_DECLARATION
Checkpatch-Ignore: MULTISTATEMENT_MACRO_USE_DO_WHILE
Checkpatch-Ignore: TRAILING_STATEMENTS
Signed-off-by: Marc Rittinghaus <marc.rittinghaus@kit.edu>
Signed-off-by: Răzvan Vîrtan <virtanrazvan@gmail.com>
Signed-off-by: cristian-vijelie <cristianvijelie@gmail.com>